// src/app/feature-modules/model/prediction.model.ts
export class Prediction {

  id: number | undefined;
  itemName: string | undefined;
  qtyOfRawMaterials: any | undefined;
  priceOfRawMaterials: any | undefined;
  totalPriceOfRawMaterials: any | undefined;
  labourCost: any | undefined;
  marketingCost: any | undefined;
  shippingCost: any | undefined;
  packagingCost: any | undefined;
  miscellaneousCost: any | undefined;
  qtyOfProduct: any | undefined;
  profitMargin: any | undefined;

}